move fixtures into separate module
